projects/citybook/requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
packaging==23.1
Pillow==9.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
urlman==2.0.1
